Add proxy login smoke coverage
Tested: ./scripts/check-lines.py

Tested: ./scripts/check-docs.py

Tested: cargo fmt --check

Tested: cargo clippy --workspace --all-targets -- -D warnings

Tested: cargo test --workspace

Tested: ./gradlew --no-daemon test shadowJar

Tested: ./scripts/verify.sh

Tested: docker compose -f docker-compose.yml -f docker-compose.verify.yml run --build --rm verify

Tested: LKJMC_MINECRAFT_SMOKE=1 ./scripts/check-minecraft-smoke.sh

Tested: LKJMC_MINECRAFT_SMOKE=1 LKJMC_MINECRAFT_PLAYER_SMOKE=1 LKJMC_STORE_TEST_DATABASE_URL=postgres://lkjmc:lkjmc-dev@127.0.0.1:5432/lkjmc ./scripts/check-minecraft-smoke.sh

Not-tested: live proxy transfer command smoke

+++ crates/lkjmc-store/src/instance.rs
@@ -110,7 +110,7 @@ pub fn allocate_port(
 ) -> Result<i32, StoreError> {
     let row = client.query_opt(
         "insert into instance_ports (port, instance_id, purpose)
-         select candidate.port, $1, $2 from generate_series($3, $4) as candidate(port)
+         select candidate.port, $1, $2 from generate_series($3::integer, $4::integer) as candidate(port)
          where not exists (select 1 from instance_ports where port = candidate.port)
          order by candidate.port limit 1 returning port",
         &[&instance_id, &purpose, &range_start, &range_end],
